requirements.txt
uvicorn==0.25.0
PyJWT==2.8.0
fastapi==0.109.0
jinja2==3.1.3
numpy==1.21.0
python-multipart==0.0.6
